+++ programs/quarry-mine/src/instructions/claim_rewards.rs
@@ -211,7 +211,7 @@ impl<'info> Validate<'info> for UserClaim<'info> {
         assert_keys_eq!(self.miner.quarry, self.quarry);
 
         // rewarder
-        assert_keys_eq!(self.quarry.rewarder_key, self.rewarder);
+        assert_keys_eq!(self.quarry.rewarder, self.rewarder);
 
         Ok(())
     }