+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `stake` to `claim` in `quarry_mine::claim_rewards`.
 - Rename `Miner.quarry_key` to `Miner.quarry` in `quarry_mine`.
+- Rename `Quarry.rewarder_key` to `Quarry.rewarder` in `quarry_mine`.
 
 ## [v4.2.1]